requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12,<4.0

